removed all of the 7 compiler warnings that CDash reported this night. This includes some fixes for good old C-files which are still in use, but were claiming since years.
There are still some more warnings around.
In general, the repository version of BACI should be
kept free of compiler warnings!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15181 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_art_net/art_net_dyn_drt.cpp
@@ -60,7 +60,7 @@ Teuchos::RCP<ART::ArtNetExplicitTimeInt> dyn_art_net_drt(bool CoupledTo3D)
   // -------------------------------------------------------------------
   // check if descretization exits
   // -------------------------------------------------------------------
-  if(DRT::Problem::Instance()->NumFields() >= genprob.numartf  && CoupledTo3D)
+  if((DRT::Problem::Instance()->NumFields() >= (unsigned) genprob.numartf)  && CoupledTo3D)
   {
     return Teuchos::null;
   }
